requirements.txt
asgiref==3.3.4
astroid==2.5.3
cffi==1.14.5
colorama==0.4.4
cryptography==3.4.7
dj-database-url==0.5.0
Django==3.2
django-environ==0.4.5
django-rest-knox==4.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
pylint==2.7.4
pylint-django==2.4.3
pylint-plugin-utils==0.6
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
wrapt==1.12.1
